requirements.txt
# =============================================================================
# Face Swap Application - Python 3.10 Compatible Requirements (FIXED)
# =============================================================================
# This file contains ONLY working dependencies for Python 3.10.x
# Install with: pip install -r requirements-python310.txt
# 
# Python Version Requirement: Python 3.10.x ONLY
# Check Python version: python --version
# 
# REMOVED PROBLEMATIC PACKAGES:
# - av (PyAV) - causes Cython compilation errors
# - Some optional dependencies that cause conflicts
# =============================================================================

# =============================================================================
# CORE SCIENTIFIC STACK - STABLE VERSIONS
# =============================================================================
numpy==1.24.4
scipy==1.10.1
scikit-image==0.20.0
scikit-learn==1.2.2
pillow==9.5.0
imageio==2.31.6
imageio-ffmpeg==0.4.9
pyyaml==6.0.1
tqdm==4.66.1
matplotlib==3.7.5
jinja2==3.1.3
psutil==5.9.8
colorama==0.4.6
tabulate==0.9.0
requests==2.31.0
typing_extensions==4.8.0

# =============================================================================
# COMPUTER VISION - TESTED VERSIONS
# =============================================================================
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78

# =============================================================================
# DEEP LEARNING STACK - PYTHON 3.10 COMPATIBLE
# =============================================================================
torch==1.13.1
torchvision==0.14.1
torchaudio==0.13.1

# AI/ML packages with fixed versions
insightface==0.7.3
onnx==1.14.1
onnxruntime-gpu==1.15.1
gfpgan==1.3.8
facexlib==0.3.0
basicsr==1.4.2

# =============================================================================
# AUDIO AND VIDEO PROCESSING (NO AV PACKAGE)
# =============================================================================
moviepy==1.0.3
soundfile==0.12.1
pydub==0.25.1
librosa==0.9.1
ffmpeg-python==0.2.0

# =============================================================================
# UI FRAMEWORK
# =============================================================================
PySide6==6.5.3

# =============================================================================
# SYSTEM UTILITIES
# =============================================================================
GPUtil==1.4.0
python-dotenv==1.0.0

# =============================================================================
# WEB FRAMEWORKS (OPTIONAL - REMOVE IF NOT NEEDED)
# =============================================================================
fastapi==0.103.2
uvicorn==0.23.2
gradio==3.48.0

# =============================================================================
# MACHINE LEARNING UTILITIES
# =============================================================================
faiss-cpu==1.7.4
tensorboard==2.14.1

# =============================================================================
# ADDITIONAL UTILITIES
# =============================================================================
json5==0.9.14
protobuf==3.20.3
cython==0.29.36
trimesh==3.22.5

# =============================================================================
# AUDIO PROCESSING (ADDITIONAL)
# =============================================================================
praat-parselmouth==0.4.3
torchcrepe==0.0.21

# =============================================================================
# MACHINE LEARNING (ADDITIONAL)
# =============================================================================
tensorboardX==2.6.2.2

# =============================================================================
# INSTALLATION INSTRUCTIONS:
# =============================================================================
# 1. Create environment: conda create -n faceswap_py310 python=3.10
# 2. Activate: conda activate faceswap_py310
# 3. Install: pip install -r requirements-python310.txt
# 
# AFTER INSTALLATION, IF YOU NEED AV PACKAGE:
# conda install -c conda-forge av
# 
# IF DLIB FAILS ON WINDOWS:
# conda install -c conda-forge dlib
# =============================================================================